+++ spec/helpers/paypal_helper.rb
@@ -1,7 +1,7 @@
 module PayPal
   def open_popup
     return window_opened_by do
-      find('.paypal-button').click
+      find(".braintree-sheet__button--paypal").click
     end
   end
 
@@ -9,12 +9,26 @@ module PayPal
     paypal_popup = open_popup
 
     within_window paypal_popup do
-      wait_for_hermes_sandbox_to_load
-      click_link("return_url")
+      login_to_paypal
+
+      click_button("confirmButtonTop", wait: 30)
     end
+
+    # can take sandbox a while to close
+    sleep 4
   end
 
-  def wait_for_hermes_sandbox_to_load
-    expect(page).to have_text("Mock Sandbox Purchase Flow")
+  def login_to_paypal
+		expect(page).to have_text("Pay with PayPal", wait: 30)
+
+		login_iframe = find("#injectedUnifiedLogin iframe")
+
+		within_frame login_iframe do
+			fill_in("email", :with => ENV["PAYPAL_USERNAME"])
+			fill_in("password", :with => ENV["PAYPAL_PASSWORD"])
+			click_button("btnLogin")
+		end
+
+		expect(page).to have_selector("#confirmButtonTop")
   end
 end