+++ src/views/payment-sheet-views/paypal-view.js
@@ -2,7 +2,9 @@
 
 var BaseView = require('../base-view');
 var paymentOptionIDs = require('../../constants').paymentOptionIDs;
-var paypal = require('braintree-web/paypal');
+var btPaypal = require('braintree-web/paypal-checkout');
+
+var DEFAULT_LOG_LEVEL = 'warn';
 
 function PayPalView() {
   BaseView.apply(this, arguments);
@@ -14,91 +16,58 @@ PayPalView.prototype = Object.create(BaseView.prototype);
 PayPalView.prototype.constructor = PayPalView;
 PayPalView.ID = PayPalView.prototype.ID = paymentOptionIDs.paypal;
 
+PayPalView.prototype.setLogLevel = function (paypal) {
+  var level = this.model.merchantConfiguration.paypal.logLevel;
+
+  paypal.setup({
+    logLevel: level || DEFAULT_LOG_LEVEL
+  });
+};
+
 PayPalView.prototype._initialize = function () {
-  this._createPayPalButton();
-  this._authInProgress = false;
+  var self = this;
+  // We wait to require paypal-checkout here in order to respect the
+  // merchant's configured log level immediately upon instantiation.
+  var paypal = require('paypal-checkout');
+
+  this.setLogLevel(paypal);
   this.model.asyncDependencyStarting();
 
-  if (!paypal.isSupported()) {
-    this.model.asyncDependencyFailed({
-      view: this.ID,
-      error: new Error(this.strings.browserNotSupported)
-    });
-    return;
-  }
+  btPaypal.create({client: this.client}, function (err, paypalInstance) {
+    var paypalCheckoutConfig;
+    var merchantConfig = self.model.merchantConfiguration.paypal;
+    var environment = self.client.getConfiguration().gatewayConfiguration.environment === 'production' ? 'production' : 'sandbox';
 
-  paypal.create({client: this.client}, function (err, paypalInstance) {
     if (err) {
-      this.model.asyncDependencyFailed({
-        view: this.ID,
+      self.model.asyncDependencyFailed({
+        view: self.ID,
         error: err
       });
       return;
     }
 
-    this.paypalInstance = paypalInstance;
+    self.paypalInstance = paypalInstance;
 
-    this.paypalButton = this.getElementById('paypal-button');
-    this.paypalButton.addEventListener('click', this._onSelect.bind(this));
+    paypalCheckoutConfig = {
+      env: environment,
+      payment: function () {
+        return paypalInstance.createPayment(merchantConfig).catch(reportError);
+      },
+      onAuthorize: function (data) {
+        return paypalInstance.tokenizePayment(data).then(function (tokenizePayload) {
+          self.model.addPaymentMethod(tokenizePayload);
+        }).catch(reportError);
+      },
+      onError: reportError
+    };
 
-    this.model.asyncDependencyReady();
-  }.bind(this));
-};
-
-PayPalView.prototype._createPayPalButton = function () {
-  var buttonContainer = this.getElementById('paypal-button');
-  var script = document.createElement('script');
-  var scriptAttrs = {
-    'data-merchant': 'braintree',
-    'data-button': 'checkout',
-    'data-button_type': 'button',
-    'data-color': 'gold',
-    'data-size': 'small'
-  };
-
-  script.src = 'https://www.paypalobjects.com/api/button.js';
-  script.async = true;
-
-  Object.keys(scriptAttrs).forEach(function (attr) {
-    script.setAttribute(attr, scriptAttrs[attr]);
+    paypal.Button.render(paypalCheckoutConfig, '[data-braintree-id="paypal-button"]').then(function () {
+      self.model.asyncDependencyReady();
+    });
   });
 
-  buttonContainer.appendChild(script);
-};
-
-PayPalView.prototype._tokenize = function () {
-  var tokenizeReturn;
-
-  this._authInProgress = true;
-
-  tokenizeReturn = this.paypalInstance.tokenize(this.model.merchantConfiguration.paypal, function (tokenizeErr, tokenizePayload) {
-    this._authInProgress = false;
-
-    if (tokenizeErr) {
-      if (tokenizeErr.code !== 'PAYPAL_POPUP_CLOSED') {
-        if (tokenizeErr.type === 'MERCHANT') {
-          console.error(tokenizeErr);
-          this.model.reportError(null);
-        } else {
-          this.model.reportError(tokenizeErr);
-        }
-      }
-      return;
-    }
-
-    this.model.addPaymentMethod(tokenizePayload);
-  }.bind(this));
-
-  this._focusFrame = tokenizeReturn.focus;
-  this.closeFrame = tokenizeReturn.close;
-};
-
-PayPalView.prototype._onSelect = function (event) {
-  event.preventDefault();
-  if (this._authInProgress) {
-    this._focusFrame();
-  } else {
-    this._tokenize();
+  function reportError(err) {
+    self.model.reportError(err);
   }
 };