+++ test/unit/dropin.js
@@ -6,7 +6,7 @@ var EventEmitter = require('../../src/lib/event-emitter');
 var analytics = require('../../src/lib/analytics');
 var fake = require('../helpers/fake');
 var hostedFields = require('braintree-web/hosted-fields');
-var paypal = require('braintree-web/paypal');
+var paypalCheckout = require('braintree-web/paypal-checkout');
 
 describe('Dropin', function () {
   beforeEach(function () {
@@ -29,7 +29,7 @@ describe('Dropin', function () {
     };
 
     this.sandbox.stub(hostedFields, 'create').yieldsAsync(null, fake.hostedFieldsInstance);
-    this.sandbox.stub(paypal, 'create').yieldsAsync(null, fake.paypalInstance);
+    this.sandbox.stub(paypalCheckout, 'create').yieldsAsync(null, fake.paypalInstance);
   });
 
   afterEach(function () {
@@ -68,7 +68,7 @@ describe('Dropin', function () {
       var hostedFieldsError = new Error('HostedFields Error');
 
       hostedFields.create.yieldsAsync(hostedFieldsError);
-      paypal.create.yieldsAsync(paypalError);
+      paypalCheckout.create.yieldsAsync(paypalError);
 
       this.sandbox.stub(analytics, 'sendEvent');
       this.dropinOptions.merchantConfiguration.paypal = {flow: 'vault'};
@@ -103,7 +103,7 @@ describe('Dropin', function () {
       var instance;
       var paypalError = new Error('PayPal Error');
 
-      paypal.create.yieldsAsync(paypalError);
+      paypalCheckout.create.yieldsAsync(paypalError);
       this.dropinOptions.merchantConfiguration.paypal = {flow: 'vault'};
 
       instance = new Dropin(this.dropinOptions);