+++ test/unit/views/payment-sheet-views/paypal-view.js
@@ -1,34 +1,51 @@
 'use strict';
+/* eslint-disable no-new */
+/* eslint-disable */
 
+var Promise = require('../../../../src/lib/promise');
 var BaseView = require('../../../../src/views/base-view');
 var DropinModel = require('../../../../src/dropin-model');
 var fake = require('../../../helpers/fake');
 var fs = require('fs');
-var PayPal = require('braintree-web/paypal');
+var PayPalCheckout = require('braintree-web/paypal-checkout');
+var paypal = require('paypal-checkout');
 var PayPalView = require('../../../../src/views/payment-sheet-views/paypal-view');
 
 var mainHTML = fs.readFileSync(__dirname + '/../../../../src/html/main.html', 'utf8');
 
+function waitForInitialize(func) {
+  setTimeout(func, 100);
+}
+
 describe('PayPalView', function () {
   beforeEach(function () {
-    var model = new DropinModel(fake.modelOptions());
+    this.model = new DropinModel(fake.modelOptions());
 
     this.div = document.createElement('div');
     this.div.innerHTML = mainHTML;
     document.body.appendChild(this.div);
     this.element = document.body.querySelector('.braintree-sheet.braintree-paypal');
 
-    model.supportedPaymentOptions = ['card', 'paypal'];
-    model.merchantConfiguration.paypal = {flow: 'vault'};
+    this.model.supportedPaymentOptions = ['card', 'paypal'];
+    this.model.merchantConfiguration.paypal = {flow: 'vault'};
+    this.sandbox.stub(this.model, 'reportError');
 
+    this.configuration = fake.configuration();
     this.paypalViewOptions = {
       strings: {},
       element: this.element,
-      model: model,
+      model: this.model,
       client: {
-        getConfiguration: fake.configuration
+        getConfiguration: this.sandbox.stub().returns(this.configuration),
+        request: this.sandbox.spy()
       }
     };
+    this.paypalInstance = {
+      createPayment: this.sandbox.stub().returns(Promise.resolve()),
+      tokenizePayment: this.sandbox.stub().returns(Promise.resolve())
+    };
+    this.sandbox.stub(PayPalCheckout, 'create').yieldsAsync(null, this.paypalInstance);
+    this.sandbox.stub(paypal.Button, 'render').returns(Promise.resolve());
   });
 
   afterEach(function () {
@@ -52,86 +69,59 @@ describe('PayPalView', function () {
   });
 
   describe('_initialize', function () {
-    beforeEach(function () {
-      this.model = new DropinModel(fake.modelOptions());
-      this.options = {
-        client: {
-          getConfiguration: fake.configuration,
-          request: this.sandbox.spy()
-        },
-        paypal: {flow: 'vault'}
-      };
-
-      this.focusStub = this.sandbox.stub();
-      this.closeStub = this.sandbox.stub();
-
-      this.tokenizeStub = this.sandbox.stub().returns({
-        focus: this.focusStub,
-        close: this.closeStub
-      });
-
-      this.paypalInstance = {tokenize: this.tokenizeStub};
-
-      this.sandbox.stub(PayPal, 'create').yields(null, this.paypalInstance);
-    });
-
-    it('starts async dependency', function () {
-      var payPalView;
-
+    it('starts async dependency', function (done) {
       this.sandbox.stub(DropinModel.prototype, 'asyncDependencyStarting');
 
-      payPalView = new PayPalView(this.paypalViewOptions);
+      new PayPalView(this.paypalViewOptions);
 
-      expect(payPalView.model.asyncDependencyStarting).to.be.calledOnce;
+      waitForInitialize(function () {
+        expect(DropinModel.prototype.asyncDependencyStarting).to.be.calledOnce;
+        done();
+      });
     });
 
-    it('notifies async dependency', function () {
+    it('notifies async dependency', function (done) {
       var payPalView;
 
       this.sandbox.stub(DropinModel.prototype, 'asyncDependencyReady');
 
       payPalView = new PayPalView(this.paypalViewOptions);
 
-      expect(payPalView.model.asyncDependencyReady).to.be.calledOnce;
-    });
-
-    it('creates a PayPal component', function () {
-      var payPalView = new PayPalView(this.paypalViewOptions);
-
-      expect(PayPal.create).to.be.calledWith(this.sandbox.match({
-        client: this.paypalViewOptions.client
-      }), this.sandbox.match.func);
-
-      expect(payPalView.paypalInstance).to.equal(this.paypalInstance);
-    });
-
-    it('calls asyncDependencyFailed with an error when PayPal is not supported', function () {
-      this.sandbox.stub(DropinModel.prototype, 'asyncDependencyFailed');
-      this.sandbox.stub(PayPal, 'isSupported').returns(false);
-
-      new PayPalView(this.paypalViewOptions); // eslint-disable-line no-new
-
-      expect(DropinModel.prototype.asyncDependencyFailed).to.be.calledOnce;
-      expect(DropinModel.prototype.asyncDependencyFailed).to.be.calledWith({
-        view: 'paypal',
-        error: new Error('Browser not supported.')
+      waitForInitialize(function () {
+        expect(payPalView.model.asyncDependencyReady).to.be.calledOnce;
+        done();
       });
     });
 
-    it('calls asyncDependencyFailed with an error when PayPal component creation fails', function () {
+    it('creates a PayPal component', function (done) {
+      var payPalView = new PayPalView(this.paypalViewOptions);
+
+      waitForInitialize(function () {
+        expect(PayPalCheckout.create).to.be.calledWith(this.sandbox.match({
+          client: this.paypalViewOptions.client
+        }), this.sandbox.match.func);
+        expect(payPalView.paypalInstance).to.equal(this.paypalInstance);
+        done();
+      }.bind(this));
+    });
+
+    it('calls asyncDependencyFailed with an error when PayPal component creation fails', function (done) {
       var fakeError = {
         code: 'A_REAL_ERROR_CODE'
       };
 
       this.sandbox.stub(DropinModel.prototype, 'asyncDependencyFailed');
-      PayPal.create.yields(fakeError);
+      PayPalCheckout.create.yieldsAsync(fakeError);
 
-      new PayPalView(this.paypalViewOptions); // eslint-disable-line no-new
+      new PayPalView(this.paypalViewOptions);
 
-      expect(DropinModel.prototype.asyncDependencyFailed).to.be.calledOnce;
-      expect(DropinModel.prototype.asyncDependencyFailed).to.be.calledWith({
-        view: 'paypal',
-        error: fakeError
+      waitForInitialize(function () {
+        expect(DropinModel.prototype.asyncDependencyFailed).to.be.calledOnce;
+        expect(DropinModel.prototype.asyncDependencyFailed).to.be.calledWith({
+          view: 'paypal',
+          error: fakeError
+        });
+        done();
       });
     });
 
@@ -141,7 +131,7 @@ describe('PayPalView', function () {
         code: 'A_REAL_ERROR_CODE'
       };
 
-      PayPal.create.yields(fakeError);
+      PayPalCheckout.create.yieldsAsync(fakeError);
 
       this.sandbox.stub(DropinModel.prototype, 'asyncDependencyStarting');
       paypalView = new PayPalView(this.paypalViewOptions);
@@ -149,136 +139,206 @@ describe('PayPalView', function () {
       expect(paypalView.model.asyncDependencyStarting).to.be.calledOnce;
     });
 
-    it('creates a PayPal button', function () {
-      var paypalButton;
+    it('calls paypal.Button.render', function (done) {
+      new PayPalView(this.paypalViewOptions);
 
-      new PayPalView(this.paypalViewOptions); // eslint-disable-line no-new
-
-      paypalButton = this.element.querySelector('[data-braintree-id="paypal-button"] script');
-
-      expect(paypalButton.getAttribute('src')).to.equal('https://www.paypalobjects.com/api/button.js');
-      expect(paypalButton.getAttribute('data-merchant')).to.equal('braintree');
-      expect(paypalButton.getAttribute('data-button')).to.equal('checkout');
-      expect(paypalButton.getAttribute('data-button_type')).to.equal('button');
-      expect(paypalButton.getAttribute('data-color')).to.equal('gold');
+      waitForInitialize(function () {
+        expect(paypal.Button.render).to.be.calledOnce;
+        expect(paypal.Button.render).to.be.calledWith(this.sandbox.match.object, '[data-braintree-id="paypal-button"]');
+        done();
+      }.bind(this));
     });
 
-    it('tokenizes when PayPal button is selected', function () {
-      var button = this.element.querySelector('[data-braintree-id="paypal-button"]');
+    it('sets paypal-checkout.js environment to production when gatewayConfiguration is production', function (done) {
+      this.configuration.gatewayConfiguration.environment = 'production';
+      new PayPalView(this.paypalViewOptions);
 
-      new PayPalView(this.paypalViewOptions); // eslint-disable-line no-new
-
-      button.click();
-
-      expect(this.paypalInstance.tokenize).to.be.calledWith(this.options.paypal);
+      waitForInitialize(function () {
+        expect(paypal.Button.render).to.be.calledWithMatch({
+          env: 'production'
+        });
+        done();
+      });
     });
 
-    it('sets a closeFrame function', function () {
-      var button = this.element.querySelector('[data-braintree-id="paypal-button"]');
+    it('sets paypal-checkout.js environment to sandbox when gatewayConfiguration is not production', function (done) {
+      this.configuration.gatewayConfiguration.environment = 'development';
+      new PayPalView(this.paypalViewOptions);
 
-      var paypalView = new PayPalView(this.paypalViewOptions);
-
-      expect(paypalView.closeFrame).to.not.exist;
-
-      button.click();
-
-      expect(paypalView.closeFrame).to.equal(this.closeStub);
+      waitForInitialize(function () {
+        expect(paypal.Button.render).to.be.calledWithMatch({
+          env: 'sandbox'
+        });
+        done();
+      });
     });
 
-    it('focuses the PayPal popup if the button is clicked after tokenization has started', function () {
-      var button = this.element.querySelector('[data-braintree-id="paypal-button"]');
+    it('calls payalInstance.createPayment with merchant config when checkout.js payment function is called', function (done) {
+      var paypalInstance = this.paypalInstance;
+      var model = this.model;
 
-      new PayPalView(this.paypalViewOptions); // eslint-disable-line no-new
+      paypal.Button.render.returns(Promise.resolve().then(function () {
+        // for some reason, this needs to be in a set timeout to grab the args from render
+        setTimeout(function () {
+          var paymentFunction = paypal.Button.render.getCall(0).args[0].payment;
 
-      button.click();
+          paymentFunction().then(function () {
+            waitForInitialize(function () {
+              expect(paypalInstance.createPayment).to.be.calledOnce;
+              expect(paypalInstance.createPayment).to.be.calledWith(model.merchantConfiguration.paypal);
+              done();
+            });
+          });
+        }, 0);
+      }));
 
-      expect(this.paypalInstance.tokenize).to.be.calledWith(this.options.paypal);
-
-      button.click();
-
-      expect(this.focusStub).to.have.been.called;
+      new PayPalView(this.paypalViewOptions);
     });
 
-    it('adds a new payment method when tokenize is successful', function () {
-      var button = this.element.querySelector('[data-braintree-id="paypal-button"]');
+    it('reports errors from createPayment', function (done) {
+      var model = this.model;
+      var error = new Error('create payment error');
 
-      this.sandbox.stub(DropinModel.prototype, 'addPaymentMethod');
-      this.tokenizeStub.yields(null, {foo: 'bar'});
+      this.paypalInstance.createPayment.returns(Promise.reject(error));
 
-      new PayPalView(this.paypalViewOptions); // eslint-disable-line no-new
+      paypal.Button.render.returns(Promise.resolve().then(function () {
+        // for some reason, this needs to be in a set timeout to grab the args from render
+        setTimeout(function () {
+          var paymentFunction = paypal.Button.render.getCall(0).args[0].payment;
 
-      button.click();
+          paymentFunction().then(function () {
+            waitForInitialize(function () {
+              expect(model.reportError).to.be.calledOnce;
+              expect(model.reportError).to.be.calledWith(error);
+              done();
+            });
+          });
+        }, 0);
+      }));
 
-      expect(this.model.addPaymentMethod).to.be.calledWith({foo: 'bar'});
+      new PayPalView(this.paypalViewOptions);
     });
 
-    it('sets _authInProgress appropriately', function () {
-      var button = this.element.querySelector('[data-braintree-id="paypal-button"]');
-
-      var paypalView = new PayPalView(this.paypalViewOptions);
-
-      expect(paypalView._authInProgress).to.be.false;
-
-      button.click();
-
-      expect(paypalView._authInProgress).to.be.true;
-
-      this.tokenizeStub.yield(null, {foo: 'bar'});
-
-      expect(paypalView._authInProgress).to.be.false;
-    });
-
-    it('reports MERCHANT errors without error payload and console errors', function () {
-      var button = this.element.querySelector('[data-braintree-id="paypal-button"]');
-      var fakeError = {type: 'MERCHANT'};
-      var paypalView = new PayPalView(this.paypalViewOptions);
-
-      this.sandbox.stub(paypalView.model, 'reportError');
-      this.sandbox.stub(console, 'error');
-      this.paypalInstance.tokenize.yields(fakeError);
-
-      button.click();
-
-      expect(paypalView.model.reportError).to.be.calledWith(null);
-      expect(console.error).to.be.calledWith(fakeError);
-    });
-
-    it('reports non-MERCHANT errors with error payload', function () {
-      var button = this.element.querySelector('[data-braintree-id="paypal-button"]');
-      var fakeError = {
-        code: 'PAYPAL_FAKE_ERROR_CODE',
-        message: 'Sorry friend',
-        type: 'NETWORK'
+    it('calls addPaymentMethod when paypal is tokenized', function (done) {
+      var paypalInstance = this.paypalInstance;
+      var model = this.model;
+      var fakePayload = {
+        foo: 'bar'
       };
-      var paypalView = new PayPalView(this.paypalViewOptions);
 
-      this.sandbox.stub(paypalView.model, 'reportError');
-      this.sandbox.stub(console, 'error');
-      this.paypalInstance.tokenize.yields(fakeError);
+      paypalInstance.tokenizePayment.returns(Promise.resolve(fakePayload));
+      this.sandbox.stub(model, 'addPaymentMethod');
 
-      button.click();
+      paypal.Button.render.returns(Promise.resolve().then(function () {
+        // for some reason, this needs to be in a set timeout to grab the args from render
+        setTimeout(function () {
+          var onAuthFunction = paypal.Button.render.getCall(0).args[0].onAuthorize;
+          var tokenizeOptions = {
+            foo: 'bar'
+          };
 
-      expect(paypalView.model.reportError).to.be.calledWith(fakeError);
-      expect(console.error).to.not.be.called;
+          onAuthFunction(tokenizeOptions);
+
+          waitForInitialize(function () {
+            expect(paypalInstance.tokenizePayment).to.be.calledOnce;
+            expect(paypalInstance.tokenizePayment).to.be.calledWith(tokenizeOptions);
+
+            expect(model.addPaymentMethod).to.be.calledOnce;
+            expect(model.addPaymentMethod).to.be.calledWith(fakePayload);
+
+            done();
+          });
+        }, 0);
+      }));
+
+      new PayPalView(this.paypalViewOptions);
     });
 
-    it('does not report PAYPAL_POPUP_CLOSED errors', function () {
-      var button = this.element.querySelector('[data-braintree-id="paypal-button"]');
-      var fakeError = {
-        code: 'PAYPAL_POPUP_CLOSED',
-        message: 'Sorry friend',
-        type: 'CUSTOMER'
+    it('reports errors from tokenizePayment', function (done) {
+      var paypalInstance = this.paypalInstance;
+      var model = this.model;
+      var error = new Error('tokenize error');
+
+      paypalInstance.tokenizePayment.returns(Promise.reject(error));
+      this.sandbox.stub(model, 'addPaymentMethod');
+
+      paypal.Button.render.returns(Promise.resolve().then(function () {
+        // for some reason, this needs to be in a set timeout to grab the args from render
+        setTimeout(function () {
+          var onAuthFunction = paypal.Button.render.getCall(0).args[0].onAuthorize;
+          var tokenizeOptions = {
+            foo: 'bar'
+          };
+
+          onAuthFunction(tokenizeOptions);
+
+          waitForInitialize(function () {
+            expect(model.reportError).to.be.calledOnce;
+            expect(model.reportError).to.be.calledWith(error);
+
+            done();
+          });
+        }, 0);
+      }));
+
+      new PayPalView(this.paypalViewOptions);
+    });
+
+    it('reports errors from paypal-checkout', function (done) {
+      var model = this.model;
+
+      paypal.Button.render.returns(Promise.resolve().then(function () {
+        // for some reason, this needs to be in a set timeout to grab the args from render
+        setTimeout(function () {
+          var onErrorFunction = paypal.Button.render.getCall(0).args[0].onError;
+          var err = new Error('Some error');
+
+          onErrorFunction(err);
+
+          waitForInitialize(function () {
+            expect(model.reportError).to.be.calledOnce;
+            expect(model.reportError).to.be.calledWith(err);
+
+            done();
+          });
+        }, 0);
+      }));
+
+      new PayPalView(this.paypalViewOptions);
+    });
+  });
+
+  describe('setLogLevel', function () {
+    beforeEach(function () {
+      this.context = {
+        model: {
+          merchantConfiguration: {
+            paypal: {}
+          }
+        }
       };
-      var paypalView = new PayPalView(this.paypalViewOptions);
+      this.fakePayPal = {
+        setup: this.sandbox.stub()
+      };
+    });
 
-      this.sandbox.stub(paypalView.model, 'reportError');
-      this.sandbox.stub(console, 'error');
-      this.paypalInstance.tokenize.yields(fakeError);
+    it('sets log level to a default value of "warn"', function () {
+      PayPalView.prototype.setLogLevel.call(this.context, this.fakePayPal);
 
-      button.click();
+      expect(this.fakePayPal.setup).to.be.calledOnce;
+      expect(this.fakePayPal.setup).to.be.calledWith({
+        logLevel: 'warn'
+      });
+    });
 
-      expect(paypalView.model.reportError).to.be.not.be.called;
-      expect(console.error).to.not.be.called;
+    it('sets log level to value set in merchant configuration', function () {
+      this.context.model.merchantConfiguration.paypal.logLevel = 'debug';
+      PayPalView.prototype.setLogLevel.call(this.context, this.fakePayPal);
+
+      expect(this.fakePayPal.setup).to.be.calledOnce;
+      expect(this.fakePayPal.setup).to.be.calledWith({
+        logLevel: 'debug'
+      });
     });
   });
 });