+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 - Add `paymentOptionPriority` option for specifying the ordering of payment options such as `card` and `paypal`
 - Add translations
 - Use version 3.11.0 of braintree-web
+- Use PayPal Checkout for PayPal View
 
 1.0.0-beta.5
 ------------